Pull in enzyme and sinon get jsx test running
