+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "bundle": "webpack -p",
     "dev": "webpack -d --watch",
-    "test": "mocha \"./test/**/*.spec.js\"",
-    "tdd": "mocha -w \"./test/**/*.spec.js\"",
+    "test": "mocha \"./test/**/*.spec.{js,jsx}\"",
+    "tdd": "mocha -w \"./test/**/*.spec.{js,jsx}\"",
     "zip": "rm -f public/bundle.map.js && zip -r -X chrome-web-store.zip . -i public\\* -i \"index.html\" -i \"manifest.json\""
   },
   "author": "",
@@ -27,8 +27,12 @@
     "babel-preset-env": "^1.6.0",
     "babel-preset-es2017": "6.24.1",
     "chai": "4.1.0",
+    "enzyme": "2.8.0",
+    "enzyme-adapter-react-15": "^1.0.1",
     "jsdom": "11.1.0",
     "mocha": "3.4.2",
+    "react-addons-test-utils": "^15.6.2",
+    "sinon": "4.0.1",
     "source-map-loader": "^0.2.1",
     "webpack": "3.4.1"
   }
