+++ test/setup.js
@@ -1,6 +1,7 @@
 const { JSDOM } = require('jsdom');
 const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
 const { window } = jsdom;
+const sinon = require('sinon');
 
 function copyProps(src, target) {
   const props = Object.getOwnPropertyNames(src)
@@ -10,6 +11,10 @@ function copyProps(src, target) {
 }
 
 global.window = window;
+global.window.localStorage = {
+  setItem: sinon.stub().returns(undefined),
+  getItem: sinon.stub().returns("{}")
+}
 global.document = window.document;
 global.navigator = {
   userAgent: 'node.js'
